+++ lib/src/api/algod/transformers/transaction_transformer.dart
@@ -3,7 +3,6 @@ import 'dart:convert';
 import 'package:algorand_dart/algorand_dart.dart';
 import 'package:algorand_dart/src/api/algod/signed_transaction_with_ad.dart';
 import 'package:algorand_dart/src/api/algod/transformers/transformers.dart';
-import 'package:algorand_dart/src/api/converters/byte_array_to_b32_converter.dart';
 import 'package:algorand_dart/src/api/converters/byte_array_to_b64_converter.dart';
 
 class TransactionTransformer
@@ -27,7 +26,7 @@ class TransactionTransformer
       sender: txn.sender?.encodedAddress ?? '',
       type: txn.type ?? '',
       note: const ByteArrayToB64Converter().fromJson(txn.note),
-      group: const ByteArrayToB32Converter().fromJson(txn.group),
+      group: const ByteArrayToB64Converter().fromJson(txn.group),
       lease: const ByteArrayToB64Converter().fromJson(txn.lease),
       rekeyTo: txn.rekeyTo?.encodedAddress,
       signature: TransactionSignature(