+++ lib/src/models/models.dart
@@ -1,12 +1,14 @@
 export 'accounts/account_model.dart';
 export 'addresses/address_model.dart';
 export 'addresses/address_role.dart';
+export 'keys/merkle_signature_verifier.dart';
 export 'keys/participation_public_key.dart';
 export 'keys/vrf_public_key.dart';
 export 'ledgers/ledger_supply.dart';
 export 'nodes/node_status_model.dart';
 export 'teal/dry_run_request_model.dart';
 export 'teal/dry_run_source_model.dart';
+export 'transactions/asset_config_parameters.dart';
 export 'transactions/atomic_transfer.dart';
 export 'transactions/on_completion_model.dart';
 export 'transactions/payment_model.dart';