+++ lib/src/utils/serializers/transaction_serializer.dart
@@ -13,9 +13,6 @@ class TransactionSerializer
       case 'pay':
         return PaymentTransaction.fromJson(data);
       case 'acfg':
-        // map apar in data
-        data.addAll(data['apar'] as Map<String, dynamic>);
-
         return AssetConfigTransaction.fromJson(data);
       case 'axfer':
         return AssetTransferTransaction.fromJson(data);